// src/app/services/files/models/file-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FileDto {
  checksum?: string;
  contentType?: string;
  duration?: number;
  entityId?: string;
  entityType?: string;
  extension?: string;
  fileSize?: number;
  folderId?: string;
  height?: number;
  humanReadableSize?: string;
  id?: string;
  isImage?: boolean;
  isPdf?: boolean;
  isPublic?: boolean;
  isVideo?: boolean;
  originalName?: string;
  path?: string;
  status?: 'ACTIVE' | 'DELETED' | 'ARCHIVED';
  storedName?: string;
  thumbnailLarge?: string;
  thumbnailMedium?: string;
  thumbnailSmall?: string;
  updatedAt?: string;
  uploadedAt?: string;
  uploadedBy?: string;
  virusScanStatus?: 'PENDING' | 'CLEAN' | 'INFECTED' | 'ERROR';
  width?: number;
}